add dynamic inertia

+++ src/BPSO/BPSOInitPop.java
@@ -43,7 +43,7 @@ public class BPSOInitPop implements InitPop{
 		for(int i = 0; i < popSize; i++){
 			for(int j = 0; j < maxVar; j++){
 				// The uniform() will return an integer value
-				realVar[i][j] = (double) StdRandom.uniform(lbound, ubound);
+				realVar[i][j] = (double) StdRandom.uniform((int) lbound, (int) ubound);
 			}
 		}
 		return realVar;
